// frontend/src/app/api/customGraphql.service.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    /** The `Upload` scalar type represents a file upload. */
    Upload: any;
};


export type Query = {
    __typename?: 'Query';
    queryUser?: Maybe<User>;
    queryUserStockWatchlists?: Maybe<Array<Maybe<StockWatchlist>>>;
};


export type QueryQueryUserArgs = {
    uid: Scalars['String'];
};


export type QueryQueryUserStockWatchlistsArgs = {
    uid: Scalars['String'];
};

export type Mutation = {
    __typename?: 'Mutation';
    updateUserData?: Maybe<User>;
    createStockWatchlist?: Maybe<StockWatchlist>;
    renameStockWatchlist?: Maybe<Scalars['Boolean']>;
    deleteWatchlist?: Maybe<Scalars['Boolean']>;
    addStockIntoStockWatchlist?: Maybe<StockDetails>;
    removeStockFromStockWatchlist?: Maybe<Scalars['Boolean']>;
};
// ...
export type User = {
    __typename?: 'User';
    uid: Scalars['ID'];
    displayName?: Maybe<Scalars['String']>;
    email?: Maybe<Scalars['String']>;
    photoURL?: Maybe<Scalars['String']>;
    providerId?: Maybe<Scalars['String']>;
    nickname?: Maybe<Scalars['String']>;
    locale?: Maybe<Scalars['String']>;
    stockWatchlist: Array<Maybe<StockWatchlist>>;
};

export type UserInput = {
    uid: Scalars['ID'];
    displayName?: Maybe<Scalars['String']>;
    email?: Maybe<Scalars['String']>;
    photoURL?: Maybe<Scalars['String']>;
    providerId?: Maybe<Scalars['String']>;
    nickname?: Maybe<Scalars['String']>;
    locale?: Maybe<Scalars['String']>;
};

export type StockWatchlist = {
    __typename?: 'StockWatchlist';
    id: Scalars['String'];
    name: Scalars['String'];
    timestamp: Scalars['Float'];
    userId: Scalars['String'];
    stocks: Array<Maybe<Scalars['String']>>;
    stocksDetails: Array<Maybe<StockDetails>>;
};

export type StockDetails = {
    __typename?: 'StockDetails';
    id: Scalars['String'];
    basicInfo?: Maybe<BasicInfo>;
    chartInfo?: Maybe<ChartInfo>;
    dividend?: Maybe<Dividend>;
    analysis?: Maybe<Analysis>;
    balanceSheet?: Maybe<BalanceSheet>;
    cashFlow?: Maybe<CashFlow>;
    overview?: Maybe<OverView>;
    perShare?: Maybe<PerShare>;
    valuation?: Maybe<Valuation>;
    financialStrength?: Maybe<FinancialStrength>;
    financialReports?: Maybe<Array<Maybe<FinancialReportNames>>>;
    financialStrengthRatio?: Maybe<FinancialStrengthRatio>;
    incomeStatement?: Maybe<IncomeStatement>;
    recommendation?: Maybe<Array<Maybe<Recommendations>>>;
    stockNewsSnippets?: Maybe<Array<Maybe<StockArticle>>>;
};
// ...
export type QueryUserQueryVariables = Exact<{
    uid: Scalars['String'];
}>;


export type QueryUserQuery = (
    { __typename?: 'Query' }
    & {
    queryUser?: Maybe<(
        { __typename?: 'User' }
        & Pick<User, 'uid' | 'displayName' | 'email' | 'photoURL' | 'nickname' | 'locale'>
        )>
}
    );

export type UpdateUserDataMutationVariables = Exact<{
    userInput: UserInput;
}>;


export type UpdateUserDataMutation = (
    { __typename?: 'Mutation' }
    & {
    updateUserData?: Maybe<(
        { __typename?: 'User' }
        & Pick<User, 'uid' | 'displayName' | 'email' | 'photoURL' | 'nickname'>
        )>
}
    );
// ...
export const QueryUserDocument = gql`
    query queryUser($uid: String!) {
        queryUser(uid: $uid) {
            uid
            displayName
            email
            photoURL
            nickname
            locale
        }
    }
`;

@Injectable({
    providedIn: 'root'
})
export class QueryUserGQL extends Apollo.Query<QueryUserQuery, QueryUserQueryVariables> {
    document = QueryUserDocument;

}

export const UpdateUserDataDocument = gql`
    mutation updateUserData($userInput: UserInput!) {
        updateUserData(user: $userInput) {
            uid
            displayName
            email
            photoURL
            nickname
        }
    }
`;

@Injectable({
    providedIn: 'root'
})
export class UpdateUserDataGQL extends Apollo.Mutation<UpdateUserDataMutation, UpdateUserDataMutationVariables> {
    document = UpdateUserDataDocument;

}